Arruma a adição de imagem no modal editar

+++ core/database/QueryBuilder.php
@@ -55,15 +55,19 @@ class QueryBuilder
 
     public function update($table, $id, $parameters, $image, $fotoAtual)
     {
-        if(isset($image)){
-            unlink($fotoAtual);
-            $pasta = "uploads/";
-            $extensao = pathinfo($image['name'], PATHINFO_EXTENSION);
-            $nomeimg = uniqid() . '.' . $extensao;
-            $caminhoimg = $pasta . basename($nomeimg);
-            move_uploaded_file($image['tmp_name'], $caminhoimg);
-            $parameters['image'] = $caminhoimg;
+        if($image && isset($image['tmp_name']) && $image['tmp_name']){
+            if(file_exists($fotoAtual)){
+                unlink($fotoAtual);
+            }
         }
+
+        $pasta = "uploads/";
+        $extensao = pathinfo($image['name'], PATHINFO_EXTENSION);
+        $nomeimg = uniqid() . '.' . $extensao;
+        $caminhoimg = $pasta . basename($nomeimg);
+        move_uploaded_file($image['tmp_name'], $caminhoimg);
+
+        $parameters['image'] = $caminhoimg;
 
         $sql = sprintf('UPDATE %s SET %s WHERE id = %s', 
             $table,
